completed homework for Week 4
